Introduce dedicated test class
